Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/content-security-policy/navigation/to-javascript-url-script-src-expected.txt

CONSOLE MESSAGE: line 1: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 1: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: line 1: Refused to execute a script because its hash, its nonce, or 'unsafe-inline' does not appear in the script-src directive of the Content Security Policy.


Harness Error (TIMEOUT), message = null

TIMEOUT <iframe src='javascript:'> blocked without 'unsafe-inline'. Test timed out
FAIL <iframe> navigated to 'javascript:' blocked without 'unsafe-inline'. assert_unreached: The CSP event should be fired in the embedding document, not in the embedee. Reached unreachable code
TIMEOUT <iframe src='...'> with 'unsafe-inline' navigated to 'javascript:' blocked in this document Test timed out
FAIL <iframe src='...'> without 'unsafe-inline' navigated to 'javascript:' blocked in this document. assert_unreached: The CSP event should be fired in the embedding document, not in the embedee. Reached unreachable code
